Move ECS Executor from __init__.py to its own file. This improves the logging because logs record the filename, and __init__.py was not a helpful name.
